src/server/requirements.txt
python-socketio==5.5.1
starlette==0.17.1
uvicorn==0.17.0
